Alpha~3 (UNTESTED)
- Added First Hit Reach check.
- Made Reach check wayy more sensitive.
- Added Vertical Block Flight check.
- Cleaned up some code.
- Removed Phase check temporarily.
- Fixed Minecraft glitch when it comes to detecting Reach.
- Replaced NoFall check with a new and improved one.
- Added Killaura (Invalid Swing) check.
- Fixed brain-fart in Killaura check that didn't allow it to flag.
- Added KillauraEvent to base.
- Added some lag checks.
- Added Killaura Direction check.
- Did some more which I probably forgot about.

+++ Basic/src/anticheat/checks/other/InvalidPackets.java
@@ -52,8 +52,8 @@ public class InvalidPackets extends Checks {
 			int verbose = 0;
 			long Time = -1L;
 			if (this.verbose.containsKey(player)) {
-				verbose = ((Integer) (this.verbose.get(player)).getKey()).intValue();
-				Time = ((Long) (this.verbose.get(player)).getValue()).longValue();
+				verbose = this.verbose.get(player).getKey().intValue();
+				Time = this.verbose.get(player).getValue().longValue();
 			}
 			verbose++;
 			if (this.verbose.containsKey(player)) {
